Refactor StatusListToken implementors to seperate files
