+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/AgentSdJwtTest.kt
@@ -24,7 +24,8 @@ import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.SD_JWT
 import at.asitplus.wallet.lib.data.CredentialPresentation.PresentationExchangePresentation
 import at.asitplus.wallet.lib.data.CredentialPresentationRequest
 import at.asitplus.wallet.lib.data.KeyBindingJws
-import at.asitplus.wallet.lib.data.StatusListToken
+import at.asitplus.wallet.lib.data.StatusListCwt
+import at.asitplus.wallet.lib.data.StatusListJwt
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatusValidationResult
 import at.asitplus.wallet.lib.data.rfc3986.toUri
 import at.asitplus.wallet.lib.extensions.sdHashInput
@@ -63,11 +64,11 @@ val AgentSdJwtTest by testSuite {
             validator = Validator(
                 tokenStatusResolver = TokenStatusResolverImpl(
                     resolveStatusListToken = {
-                        if (Random.nextBoolean()) StatusListToken.StatusListJwt(
+                        if (Random.nextBoolean()) StatusListJwt(
                             statusListIssuer.issueStatusListJwt(),
                             resolvedAt = Clock.System.now()
                         ) else {
-                            StatusListToken.StatusListCwt(
+                            StatusListCwt(
                                 statusListIssuer.issueStatusListCwt(),
                                 resolvedAt = Clock.System.now(),
                             )
