+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/ValidatorVpTest.kt
@@ -11,7 +11,8 @@ import at.asitplus.wallet.lib.data.ConstantIndex
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.PLAIN_JWT
 import at.asitplus.wallet.lib.data.CredentialPresentation.PresentationExchangePresentation
 import at.asitplus.wallet.lib.data.CredentialPresentationRequest
-import at.asitplus.wallet.lib.data.StatusListToken
+import at.asitplus.wallet.lib.data.StatusListCwt
+import at.asitplus.wallet.lib.data.StatusListJwt
 import at.asitplus.wallet.lib.data.VerifiablePresentation
 import at.asitplus.wallet.lib.data.VerifiablePresentationJws
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
@@ -63,11 +64,11 @@ val ValidatorVpTest by testSuite {
             validator = Validator(
                 tokenStatusResolver = TokenStatusResolverImpl(
                     resolveStatusListToken = {
-                        if (Random.nextBoolean()) StatusListToken.StatusListJwt(
+                        if (Random.nextBoolean()) StatusListJwt(
                             statusListIssuer.issueStatusListJwt(),
                             resolvedAt = Clock.System.now()
                         ) else {
-                            StatusListToken.StatusListCwt(
+                            StatusListCwt(
                                 statusListIssuer.issueStatusListCwt(),
                                 resolvedAt = Clock.System.now()
                             )
